% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{l_time_domain}
\alias{l_time_domain}
\title{Laplace Transforms}
\format{An object of class \code{character} of length 37.}
\usage{
l_time_domain
}
\description{
The datasets \code{l_time_domain} and \code{l_laplace_s_domain} provide Laplace
transforms for many common functions of a single variable. Taken as a pair
of equal-length vectors, the LaTeX formulas within each index position
represent a common function, where any instance within the \code{l_time_domain}
vector that function in the time domain, and, the corresponding Laplace
s-domain function exists in the \code{l_laplace_s_domain} vector.
}
\keyword{datasets}
